Refactors sky rendering
