+++ packages/rewild-renderer/lib/core/AtmosphereSkybox.ts
@@ -1,43 +1,53 @@
 import { Renderer } from '..';
 import { BoxGeometryFactory } from '../geometry/BoxGeometryFactory';
-import { AtmosphereCubeMaterial3 } from '../materials/AtmosphereCubeMaterial3';
+import { Geometry } from '../geometry/Geometry';
+import { SkyRenderer } from '../materials/sky/SkyRenderer';
 import { Camera } from './Camera';
-import { Mesh } from './Mesh';
+import { Transform } from './Transform';
 
 export class AtmosphereSkybox {
-  mesh: Mesh;
-  material: AtmosphereCubeMaterial3;
+  transform: Transform;
+  geometry: Geometry;
+  skyRenderer: SkyRenderer;
   initialized: boolean;
 
   constructor() {
     this.initialized = false;
-    this.material = new AtmosphereCubeMaterial3();
-    this.mesh = new Mesh(BoxGeometryFactory.new(), this.material);
-    this.mesh.transform.scale.set(1000, 1000, 1000);
-    this.mesh.transform.updateMatrixWorld();
+    this.geometry = BoxGeometryFactory.new();
+    this.transform = new Transform();
+    this.skyRenderer = new SkyRenderer();
+    this.transform.scale.set(1000, 1000, 1000);
+    this.transform.updateMatrixWorld();
   }
 
   render(renderer: Renderer, pass: GPURenderPassEncoder, camera: Camera) {
-    this.mesh.transform.modelViewMatrix.multiplyMatrices(
-      camera.matrixWorldInverse,
-      this.mesh.transform.matrixWorld
-    );
-    this.mesh.transform.normalMatrix.getNormalMatrix(
-      this.mesh.transform.modelViewMatrix
+    this.transform.position.set(
+      camera.transform.position.x,
+      camera.transform.position.y,
+      camera.transform.position.z
     );
 
+    this.transform.updateMatrixWorld();
+
+    this.transform.modelViewMatrix.multiplyMatrices(
+      camera.matrixWorldInverse,
+      this.transform.matrixWorld
+    );
+
+    this.transform.normalMatrix.getNormalMatrix(this.transform.modelViewMatrix);
+
     if (!this.initialized) {
-      this.mesh.geometry.build(renderer.device);
-      this.material.init(renderer);
+      this.geometry.build(renderer.device);
+      this.skyRenderer.init(renderer);
       this.initialized = true;
     }
 
-    this.material.render(
+    this.skyRenderer.render(
       renderer,
       pass,
       camera,
-      [this.mesh],
-      this.mesh.geometry
+      this.transform,
+      this.geometry
     );
   }
 }
