+++ packages/rewild-renderer/lib/materials/sky/AtmosphereMaterial.ts
@@ -1,15 +1,15 @@
-import { Geometry } from '../geometry/Geometry';
-import { IMaterialPass } from './IMaterialPass';
+import { Geometry } from '../../geometry/Geometry';
+import { IMaterialPass } from '../IMaterialPass';
 import shader from '../shaders/atmospherePlane.wgsl';
-import { Renderer } from '..';
-import { ProjModelView } from './uniforms/ProjModelView';
-import { PerMeshTracker } from './PerMeshTracker';
-import { SharedUniformsTracker } from './SharedUniformsTracker';
-import { Mesh } from '../core/Mesh';
-import { Camera } from '../core/Camera';
-import { Diffuse } from './uniforms/Diffuse';
-import { samplerManager } from '../textures/SamplerManager';
-import { textureManager } from '../textures/TextureManager';
+import { Renderer } from '../..';
+import { ProjModelView } from '../uniforms/ProjModelView';
+import { PerMeshTracker } from '../PerMeshTracker';
+import { SharedUniformsTracker } from '../SharedUniformsTracker';
+import { Mesh } from '../../core/Mesh';
+import { Camera } from '../../core/Camera';
+import { Diffuse } from '../uniforms/Diffuse';
+import { samplerManager } from '../../textures/SamplerManager';
+import { textureManager } from '../../textures/TextureManager';
 import { Matrix4 } from 'rewild-common';
 
 const sharedBindgroupIndex = 1;
