+++ packages/rewild-renderer/lib/shaders/atmosphereBloomPass.wgsl
@@ -23,6 +23,7 @@ var<uniform> object: ObjectStruct;
   var sum = vec4f(0.0);
   let NUM_SAMPLES = 10.0;
   let phiOffset = hash(dot(fragCoord.xy, vec2(1.12,2.251)) + time * 0.0001);
+
   for(var i: i32 = 0; i < i32(NUM_SAMPLES); i++)
   {
       let r = blurRadius * f32(i) / NUM_SAMPLES;
@@ -32,11 +33,11 @@ var<uniform> object: ObjectStruct;
       sum += textureSampleLevel( ourTexture, ourSampler, uv, 0.0);
   }
 
-  let BLOOM_AMOUNT = 0.1; 
+  let BLOOM_AMOUNT = 0.01; 
   let originalColor = textureSampleLevel( ourTexture, ourSampler, uv, 0.0);
   sum = vec4f( mix(originalColor.xyz, sum.xyz / NUM_SAMPLES, BLOOM_AMOUNT), originalColor.w );
 
-  let exposure = 0.06;// * ( 1.0 + 0.2 * sin( 0.5 ) * sin( 1.8  ));
+  let exposure = 0.09;// * ( 1.0 + 0.2 * sin( 0.5 ) * sin( 1.8  ));
   return vec4f(tonemapACES( exposure * sum.xyz ), sum.w); 
 }
