+++ packages/rewild-renderer/lib/shaders/atmosphereTAAPass.wgsl
@@ -54,7 +54,7 @@ const offsets = array<vec2i, 8>(
     
   history = clamp(history, colorMin, colorMax);
   
-	return vec4f(YCoCgToRGB(mix(newFrame, history, 0.80)), newFrameSample.a);
+	return vec4f(YCoCgToRGB(mix(newFrame, history, 0.60)), newFrameSample.a);
 }
